Update Driver License Controller and services

+++ backend/src/main/java/com/quince/rentingapp/repository/DriverLicenseRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 
 public interface DriverLicenseRepository extends JpaRepository<DriverLicense, Long> {
 
-   Optional<DriverLicense> findByTCno(int TC);
+   Optional<DriverLicense> findByTCno(long TC);
    Optional<DriverLicense> findByDocumentNo(long documentNo);
    List<DriverLicense> findByIntern(boolean intern);
    List<DriverLicense> findByOffice(String office);
